requirements.txt
orjson==3.8.*
PyJWT==2.7.*
requests~=2.28
dataclasses==0.6
validators>=0.22.0
httpx==0.24.*
PyYAML>=6.0
py-easy-logger>=0.0.5
